workshop/demo-ocr/image-to-text/requirements.txt
openai

Pillow
pytesseract
easyocr